// x-pack/plugins/security_solution/common/generated_schema/find_rules/find_rules_response_schema.gen.ts
/*
 * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
 * or more contributor license agreements. Licensed under the Elastic License
 * 2.0; you may not use this file except in compliance with the Elastic License
 * 2.0.
 */

import { z } from 'zod';
import * as CommonSchema from '../common_schema.gen';

/*
 * NOTICE: Do not edit this file manually.
 * This file is automatically generated by the OpenAPI Generator `yarn openapi:generate`.
 */

export const FindRulesResponse = z.object({
  page: z.number().optional(),
  perPage: z.number().optional(),
  total: z.number().optional(),
  data: z.array(CommonSchema.Rule).optional(),
});
export type FindRulesResponse = z.infer<typeof FindRulesResponse>;