// x-pack/plugins/security_solution/common/generated_schema/find_rules/find_rules_request_schema.gen.ts
/*
 * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
 * or more contributor license agreements. Licensed under the Elastic License
 * 2.0; you may not use this file except in compliance with the Elastic License
 * 2.0.
 */

import { z } from 'zod';
import * as CommonSchema from '../common_schema.gen';

/*
 * NOTICE: Do not edit this file manually.
 * This file is automatically generated by the OpenAPI Generator `yarn openapi:generate`.
 */

export const FindRulesRequestQuery = z.object({
  /**
   * The page number to return.
   */
  page: z.number().optional(),
  /**
   * The number of rules to return per page.
   */
  per_page: z.number().optional(),
  /**
   * Determines which field is used to sort the results.
   */
  sort_field: CommonSchema.FindRulesSortField.optional(),
  /**
   * Determines the sort order, which can be `desc` or `asc`.
   */
  sort_order: CommonSchema.SortOrder.optional(),
  /**
   * Filters the returned results according to the value of the specified field, using the `alert.attributes.<field name>:<field value>` syntax.
   */
  filter: z.string().optional(),
  /**
   * An array of fields to include in the results.
   */
  fields: z.array(z.string()).optional(),
});
export type FindRulesRequestQuery = z.infer<typeof FindRulesRequestQuery>;
export type FindRulesRequestQueryInput = z.input<typeof FindRulesRequestQuery>;

export const FindRulesRequestParams = z.undefined();
export type FindRulesRequestParams = z.infer<typeof FindRulesRequestParams>;
export type FindRulesRequestParamsInput = z.input<typeof FindRulesRequestParams>;

export const FindRulesRequestBody = z.undefined();
export type FindRulesRequestBody = z.infer<typeof FindRulesRequestBody>;
export type FindRulesRequestBodyInput = z.input<typeof FindRulesRequestBody>;